+++ test/govspeak_images_bang_test.rb
@@ -7,12 +7,13 @@ class GovspeakImagesBangTest < Minitest::Test
   include GovspeakTestHelper
 
   class Image
-    attr_reader :alt_text, :url, :caption
+    attr_reader :alt_text, :url, :caption, :credit
 
     def initialize(attrs = {})
       @alt_text = attrs[:alt_text] || "my alt"
       @url = attrs[:url] || "http://example.com/image.jpg"
       @caption = attrs[:caption]
+      @credit = attrs[:credit]
     end
   end
 
@@ -51,7 +52,7 @@ class GovspeakImagesBangTest < Minitest::Test
       assert_html_output(
         %{<figure class="image embedded">} +
         %{<div class="img"><img src="http://example.com/image.jpg" alt="my alt"></div>\n} +
-        %{<figcaption>My Caption &amp; so on</figcaption>} +
+        %{<figcaption><p>My Caption &amp; so on</p></figcaption>} +
         %{</figure>}
       )
     end
@@ -62,6 +63,41 @@ class GovspeakImagesBangTest < Minitest::Test
       assert_html_output(
         %{<figure class="image embedded">} +
         %{<div class="img"><img src="http://example.com/image.jpg" alt="my alt"></div>} +
+        %{</figure>}
+      )
+    end
+  end
+
+  test "¡¡n syntax adds image credit if given" do
+    given_govspeak "!!1", images: [Image.new(credit: 'My Credit & so on')] do
+      assert_html_output(
+        %{<figure class="image embedded">} +
+        %{<div class="img"><img src="http://example.com/image.jpg" alt="my alt"></div>\n} +
+        %{<figcaption><p>Image credit: My Credit &amp; so on</p></figcaption>} +
+        %{</figure>}
+      )
+    end
+  end
+
+  test "!!n syntax ignores a blank credit" do
+    given_govspeak "!!1", images: [Image.new(credit: '  ')] do
+      assert_html_output(
+        %{<figure class="image embedded">} +
+        %{<div class="img"><img src="http://example.com/image.jpg" alt="my alt"></div>} +
+        %{</figure>}
+      )
+    end
+  end
+
+  test "!!n syntax adds image caption and credit if given" do
+    given_govspeak "!!1", images: [Image.new(caption: 'My Caption & so on', credit: 'My Credit & so on')] do
+      assert_html_output(
+        %{<figure class="image embedded">} +
+        %{<div class="img"><img src="http://example.com/image.jpg" alt="my alt"></div>\n} +
+        %{<figcaption>} +
+          %{<p>My Caption &amp; so on</p>\n} +
+          %{<p>Image credit: My Credit &amp; so on</p>} +
+        %{</figcaption>} +
         %{</figure>}
       )
     end