+++ test/govspeak_images_test.rb
@@ -43,7 +43,7 @@ class GovspeakImagesTest < Minitest::Test
       assert_html_output(
         %{<figure class="image embedded">} +
         %{<div class="img"><img src="http://example.com/image.jpg" alt="my alt"></div>\n} +
-        %{<figcaption>My Caption &amp; so on</figcaption>} +
+        %{<figcaption><p>My Caption &amp; so on</p></figcaption>} +
         %{</figure>}
       )
     end
@@ -54,6 +54,41 @@ class GovspeakImagesTest < Minitest::Test
       assert_html_output(
         %{<figure class="image embedded">} +
         %{<div class="img"><img src="http://example.com/image.jpg" alt="my alt"></div>} +
+        %{</figure>}
+      )
+    end
+  end
+
+  test "Image:image-id syntax adds image credit if given" do
+    given_govspeak "[Image:image-id]", images: [build_image(credit: 'My Credit & so on')] do
+      assert_html_output(
+        %{<figure class="image embedded">} +
+        %{<div class="img"><img src="http://example.com/image.jpg" alt="my alt"></div>\n} +
+        %{<figcaption><p>Image credit: My Credit &amp; so on</p></figcaption>} +
+        %{</figure>}
+      )
+    end
+  end
+
+  test "Image:image-id syntax ignores a blank credit" do
+    given_govspeak "[Image:image-id]", images: [build_image(credit: '  ')] do
+      assert_html_output(
+        %{<figure class="image embedded">} +
+        %{<div class="img"><img src="http://example.com/image.jpg" alt="my alt"></div>} +
+        %{</figure>}
+      )
+    end
+  end
+
+  test "Image:image-id syntax adds image caption and credit if given" do
+    given_govspeak "[Image:image-id]", images: [build_image(caption: 'My Caption & so on', credit: 'My Credit & so on')] do
+      assert_html_output(
+        %{<figure class="image embedded">} +
+        %{<div class="img"><img src="http://example.com/image.jpg" alt="my alt"></div>\n} +
+        %{<figcaption>} +
+          %{<p>My Caption &amp; so on</p>\n} +
+          %{<p>Image credit: My Credit &amp; so on</p>} +
+        %{</figcaption>} +
         %{</figure>}
       )
     end